Don't update chunks when rendering lights
Due to frustum updates now being asynchronous, the updates started
while rendering from a light's perspective can "bleed" into the main
camera view and vice versa, causing chunks to flicker because they
are not visible from the other perspective.

The updateChunks mixin is just there in case that method could cause
a similar problem, it should not be necessary to fix this bug.